Metrics are now Send and Sync
So I prematurely optimized. I shouldn't have done that. The raw
UnsafeCell is optimal, for sure, but it just doesn't matter. The
uncontended impact of Mutex is so small the benchmark struggles
to discern between the 2 strategies. There are perhaps 40
nanoseconds separating uncontended Mutex from UnsafeCell on an
m1 macbook pro. Probably it's even closer on Linux with rust
1.62 and up.

This should ease integration with Tokio-based applications.
You should still avoid contending on Metrics objects but who am
I to judge?