add tdigest for goodmetrics downstreams
this is a minor breaking change. You now have to choose which kind
of aggregation you will use in the aggregating sink. It's expected
to be a trivial change for users, with ::Histogram retaining the
existing behavior. If you're doing something new with Goodmetrics
protocol you should use the default tdigest unless you have a good
reason to use histogram instead.

copied the tdigest library so I could access some internals for
quicker recording and serialization.