remove protoc from goodmetrics build dependencies
Adds an unpublished sidecar workspace project that tries to watch
the proto directory and generate protos into the main library.
In this way the main library can use the generated source directly.

This should make protoc unnecessary in user applications (at least,
on account of goodmetrics anyway). The publish workflow will ensure
that there aren't any locally modified files, so if a proto is changed
and the generated source doesn't accompany it, PR will fail.