improve contended latency
The AlwaysNewMetricsAllocator is faster, by a wide margin, in
contended scenarios. By leaning into it a little, this pr gets
nearly 18x more throughput with 8 cores (on a macbook), and still
reaches 25% more throughput at 20% lower latency with 1 core.

Broadly, there are 2 big performance increases in here:

1) Moving metrics aggregation into a separate thread, via mpsc. This
removed a Mutex that all Metrics drop()s had to go through. Now they
just go into an mpsc SyncSender.

This change made it possible to change the AggregatingSink to just
be an Aggregator. It is supposed to take a thread now, and just marshal
metrics through aggregation. The threaded context is now over `self`
as well, which means there is no need for a Mutex. Since we have
`self` and SC from mpsc, we can blindly exploit `&mut self`. This
makes metrics aggregation non-blocking and very lightweight.

By moving metrics aggregation to a dedicated thread and removing the
Mutex from the drop path, direct consumers of the record_scope() api
should see significantly lower latency and accordingly higher
throughput.

2) Removing reference requirement from MetricsAllocator. This was an
experiment, but it made a huge difference to record_scope() users.
After the improvements in 1), this further refined another 10-20%,
depending on contention. 10% would be worth the simplification, but
in single-threaded use cases it improves as much as 20%. The cost
here is that it is likely going to cause some churn.

These changes together are breaking. I may remove the pooled allocator
in the future, but I'll leave it for the time being.