Add gauges
Gauges are StatisticSets that are periodically retrieved. There is
no additional overhead for a Gauge beyond a few Relaxed memory
order atomic operations. Under high contention on a Macbook the
cost per observation of a number is around 0.25µs.

Metrics background tasks are getting to be more onerous. I will
need to simplify the execution model at some point.