add arc_allocator
A new caching allocator, like the pooled_metrics_allocator, the
arc_allocator relies on Arc instead of the borrow checker. While
structural borrow checking is definitely better than Arc, in
practice Metrics pipelines are complex enough that the lifecycles
of Metrics references downstream are massively easier to work with
when they are owned. In particular, the StreamSink is hard to use
while satisfying lifetime requirements. It is expected that in
practice the arc_allocator will not be much worse than the
pooled_metrics_allocator.

arc_allocator uses an optimistic 8-slot concurrency model. It tries
to avoid colliding locks for checking out and returning objects.
This means it might allocate extra Metrics that a strict policy
wouldn't, but synchronization is considered more likely to be
problematic than an extra Metrics object here and there.

Make sure you set your cache size to something appropriate for your
use case. The minimum is 16*8 = 128. Powers of 2 will get you accurate
buffer sizes; anything else will not.