switch streamsink to std::sync::mpsc (#12)
The tokio::sync::mpsc costs a lot of time when doing a cross-runtime
wake. A normal way to use metrics is to split the reporting thread(s)
from the service/recording threads. It's natural to put each in their
respective runtimes. That's not a good configuration for a tokio mpsc,
and it measures to be way too expensive in at least 1 service.

By contrast, the std channel does a good job across threads. The
main drawback here is that there's no await, so the Aggregator has
to poll. Probably it will be fine for expected deployment scenarios,
but of course, pr's welcome.