Add support for dimensions for gauges (#13)
Add support for dimensioned gauges. Dimensions are added to a gauge
when it is created. Gauges are stored in a GaugeGroup by their
DimensionPosition.

`GaugeDimensions` are a group of dimensions that can be built up, and
then are added to the gauge when it is created, which consumes the
`GaugeDimensions`.

---------

Co-authored-by: kvc0 <3454741+kvc0@users.noreply.github.com>