add histogram gauge, switch to gauge handles
Arc<Gauge> is annoying, and it limits the convenience of different
types of gauge. A histogram, for example, has a natural "time" function
that should return a scope guard. It's less clear that is useful for
Sum. So gauges how use cheaply clonable handles that make their type
clearer.

Histogram gauge now uses the exponential_histogram crate.